Related work items: #13953372, #14681942

+++ src/inc/ZipFileStream.hpp
@@ -19,8 +19,8 @@ namespace xPlat {
             std::string contentType,
             IxPlatFactory* factory,
             bool isCompressed,
-            std::uint32_t offset,
-            std::uint32_t size,
+            std::uint64_t offset,
+            std::uint64_t size,
             IStream* stream
         ) : m_isCompressed(isCompressed), RangeStream(offset, size, stream), m_name(name), m_contentType(contentType), m_factory(factory)
         {
